+++ test/OpenTelemetry.OpAmp.Client.Tests/OpAmpClientTests.cs
@@ -1,7 +1,9 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+using OpAmp.Proto.V1;
 using OpenTelemetry.OpAmp.Client.Internal.Services.Heartbeat;
+using OpenTelemetry.OpAmp.Client.Settings;
 using OpenTelemetry.OpAmp.Client.Tests.Mocks;
 using OpenTelemetry.OpAmp.Client.Tests.Tools;
 using Xunit;
@@ -72,6 +74,82 @@ public class OpAmpClientTests
         static ulong GetCurrentTimeInNanoseconds()
         {
             return (ulong)DateTimeOffset.UtcNow.ToUnixTimeMilliseconds() * 1_000_000; // Convert to nanoseconds
+        }
+    }
+
+    [Fact]
+    public async Task DoesNotEmitHeartbeat_WhenDisabled()
+    {
+        using var opAmpServer = new OpAmpFakeHttpServer(false);
+        var opAmpEndpoint = opAmpServer.Endpoint;
+
+        using var mockListener = new MockListener();
+        using var client = new OpAmpClient(o =>
+        {
+            o.ServerUrl = opAmpEndpoint;
+            o.Heartbeat.IsEnabled = false;
+        });
+        client.Subscribe(mockListener);
+
+        await client.StartAsync();
+
+        mockListener.WaitForMessages(TimeSpan.FromSeconds(2));
+
+        var frames = opAmpServer.GetFrames();
+
+        // Only the identification message should be received
+        Assert.Single(frames);
+        Assert.Single(mockListener.Messages);
+
+        await client.StopAsync();
+    }
+
+    [Theory]
+    [ClassData(typeof(CapabilityTestData))]
+    internal async Task SendsExpectedCapabilities(
+        Action<OpAmpClientSettings> configure,
+        IEnumerable<AgentCapabilities> expectedCapabilities,
+        IEnumerable<AgentCapabilities> notExpectedCapabilities)
+    {
+        using var opAmpServer = new OpAmpFakeHttpServer(false);
+        var opAmpEndpoint = opAmpServer.Endpoint;
+        configure += o => o.ServerUrl = opAmpEndpoint;
+
+        using var mockListener = new MockListener();
+        using var client = new OpAmpClient(configure);
+        client.Subscribe(mockListener);
+
+        await client.StartAsync();
+
+        var frames = opAmpServer.GetFrames();
+
+        Assert.True(frames.Count >= 1, "Expecting at least one server frame.");
+
+        var identificationFrame = frames[0];
+        var capabilities = (AgentCapabilities)identificationFrame.Capabilities;
+
+        foreach (var expectedCapability in expectedCapabilities)
+        {
+            Assert.True(capabilities.HasFlag(expectedCapability), $"Expected capabilities to include {expectedCapability}.");
+        }
+
+        foreach (var notExpectedCapability in notExpectedCapabilities)
+        {
+            Assert.False(capabilities.HasFlag(notExpectedCapability), $"Expected capabilities not to include {notExpectedCapability}.");
+        }
+
+        await client.StopAsync();
+    }
+
+    internal class CapabilityTestData
+        : TheoryData<Action<OpAmpClientSettings>, IEnumerable<AgentCapabilities>, IEnumerable<AgentCapabilities>>
+    {
+        public CapabilityTestData()
+        {
+            this.Add(o => o.Heartbeat.IsEnabled = false, [], [AgentCapabilities.ReportsHeartbeat, AgentCapabilities.ReportsHealth]);
+            this.Add(o => o.Heartbeat.IsEnabled = true, [AgentCapabilities.ReportsHeartbeat, AgentCapabilities.ReportsHealth], []);
+            this.Add(o => o.RemoteConfiguration.AcceptsRemoteConfig = true, [AgentCapabilities.AcceptsRemoteConfig], []);
+            this.Add(o => o.RemoteConfiguration.AcceptsRemoteConfig = false, [], [AgentCapabilities.AcceptsRemoteConfig]);
         }
     }
 }