+++ src/OpenTelemetry.OpAmp.Client/CHANGELOG.md
@@ -15,6 +15,9 @@
 * Expose public `RemoteConfigMessage`.
   ([#3614](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3614))
 
+* Add settings for remote configuration and update advertised capabilities.
+  (TODO)
+
 ## 0.1.0-alpha.3
 
 Released 2025-Nov-13